+++ kitty/config_data.py
@@ -456,6 +456,9 @@ o('rectangle_select_modifiers', 'ctrl+alt', option_type=to_modifiers, long_text=
 The modifiers to use rectangular selection (i.e. to select text in a
 rectangular block with the mouse)'''))
 
+o('terminal_select_modifiers', 'shift', option_type=to_modifiers, long_text=_('''
+The modifiers to override mouse selection even when a terminal application has grabbed the mouse'''))
+
 o('select_by_word_characters', ':@-./_~?&=%+#', long_text=_('''
 Characters considered part of a word when double clicking. In addition to these characters
 any character that is marked as an alpha-numeric character in the unicode