+++ kitty/state.c
@@ -384,6 +384,7 @@ PYWRAP1(set_options) {
     S(touch_scroll_multiplier, PyFloat_AsDouble);
     S(open_url_modifiers, convert_mods);
     S(rectangle_select_modifiers, convert_mods);
+    S(terminal_select_modifiers, convert_mods);
     S(click_interval, PyFloat_AsDouble);
     S(resize_debounce_time, PyFloat_AsDouble);
     S(url_color, color_as_int);